// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Thu, 06 Jul 2023 15:49:23 GMT
 */

/**
 * @enum
 * @readonly
 * @description AlienWorlds Action Names
 */
export enum AlienWorldsActionName {
    Addvesting = 'addvesting',
    Burn = 'burn',
    Close = 'close',
    Create = 'create',
    Issue = 'issue',
    Open = 'open',
    Transfer = 'transfer',
};